requirements.txt
ngrok
uvicorn
python-dotenv
fastapi
pydantic
numpy
keras
scikit-learn